requirements.txt
Django~=5.2.2
django-datetime-utc~=1.0.5
django-simple-history~=3.8.0
djangorestframework~=3.15.2
drf-yasg~=1.21.10
django-filter
psycopg[binary]